Proposed changes to __print__ and __repr__
I’m not pleased with the AnnoSpan version of this. I want to include `metadata` but it gives errors sometimes; I think changing AnnoSpan will fix this somewhat.

For AnnoTier, I changed it so that __repr__ is a terse representation, and __str__ is more verbose and prints contained spans line by line.