Remove proxy command line flags but keep warning message
Since it suffices to set the proxy via the shell, it is not necessary to
add command line flags for proxy settings. This way, the number of lines
of code is reduced. However, keeping the warning might be helpful for
users that do not initially know why the import might be failing.