Rewrite windows_max_int() to calculate mid correctly, return it, and shorten function
As pointed out by Nathan, it is more senseful that windows_max_int() returns mid since mid
is tested on whether it produces an overflow error or not.
Furthermore, Nathan pointed out that mid was calculated incorrectly. This commit corrects
the mistake.
Also, instead of calculating diff to determine the stopping criterion, I check
whether the difference of top to lower limit equals 1 to make code more readable.